requirements.txt
Flask==3.0.0
pymongo==4.6.0
reportlab==4.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
pytz==2024.1
flask-cors==4.0.0
PyJWT==2.8.0
